Change int to double type of weight in Nutrition Core model

+++ app/src/main/java/com/app/dmitryteplyakov/sportdiary/Core/Nutrition/Nutrition.java
@@ -66,7 +66,7 @@ public class Nutrition {
         return mWeight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         mWeight = weight;
     }
